+++ packages/design-system/src/atoms/Badge/Badge.interface.ts
@@ -1,0 +1,26 @@
+import { BaseHTMLAttributes } from 'react';
+
+export enum BadgeSize {
+    ExtraSmall = 'xs',
+    Small = 'sm',
+    Medium = 'md',
+    Large = 'lg'
+}
+
+export enum BadgeColor {
+    Red = 'red',
+    Blue = 'blue'
+}
+
+export enum BadgePosition {
+    Relative = 'relative',
+    TopRight = 'top-right',
+    TopLeft = 'top-left'
+}
+
+export interface BadgeProps extends BaseHTMLAttributes<HTMLDivElement>{
+    size?: `${BadgeSize}`,
+    color?: `${BadgeColor}`,
+    info?: string,
+    position?: `${BadgePosition}`
+}