+++ packages/design-system/src/atoms/Badge/Badge.stories.tsx
@@ -1,0 +1,19 @@
+import type { ComponentStory } from '@storybook/react';
+import Badge from './Badge';
+
+
+export const helperComponentStyle = {
+    height: 24,
+    width: 36,
+    backgroundColor: 'cyan',
+    borderRadius: 6,
+    position: 'relative',
+    display: 'flex',
+    alignItems: 'center',
+  }
+
+export const Template: ComponentStory<typeof Badge> = (args) => <Badge {...args} />;
+
+export const args = {
+    info: '8'
+}